LayoutTests/http/tests/security/contentSecurityPolicy/script-src-self-blocked-03-expected.txt

CONSOLE MESSAGE: Refused to load script from 'https://127.0.0.1:8443/security/contentSecurityPolicy/resources/script.js' because of Content-Security-Policy.



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS
